提交

+++ klinsurance-core/src/main/java/com/ht/klinsurance/loss/model/LossItem.java
@@ -14,7 +14,7 @@ public class LossItem {
 
     private String lossItemName;
 
-    private Integer lossNumber;
+    private BigDecimal lossNumber;
 
     private BigDecimal length;
